test(workflow): backend/tests/domain/workflow/ に機能単位で分割 (Norman §file-split)
ノーマン指摘の test_workflow.py 1083 行 (500 行ルール超過) を機能単位で 11 ファイルに分割。
各クラス群を独立した test_*.py に切り出し、grep / failure 報告での追跡性を向上。

新構造 (各ファイル < 200 行):
- test_construction.py (TestWorkflowConstruction / TestWorkflowName)
- test_dag_invariants.py (TestEntryStageId / TestUnreachableStage / TestSinkStage /
  TestTransitionDeterminism / TestExternalReviewNotify / TestRequiredRoleNonEmpty /
  TestTransitionRefIntegrity / TestBFSCycleSafety)
- test_mutators.py (TestAddStage / TestAddStageCapacity / TestAddTransition /
  TestAddTransitionCapacity / TestRemoveStage / TestPreValidateRollback)
- test_from_dict.py (TestFromDict)
- test_frozen_extra.py (TestFrozenContract / TestExtraForbid)
- test_notify_channel_ssrf.py (TestNotifyChannelSSRF)
- test_notify_channel_kind.py (TestNotifyChannelKindMVP)
- test_notify_channel_masking.py (TestNotifyChannelMasking)
- test_helpers_independence.py (TestHelperIndependence)
- test_integration.py (TestWorkflowLifecycleIntegration)
- test_value_objects.py (旧 test_value_objects_workflow.py の内容)

実行結果 (再構造化後):
- pytest: 160 passed in 0.31s (件数維持、全件緑)
- coverage: 99% 維持 (empire 100% / workflow 99% / value_objects 97% / exceptions 100%)
- ruff format/check: clean
- pyright strict: 0 errors

リーナスの workflow.py 分割 + _validate_transition_id_unique 追加 (スティーブ指摘) の
完了後に対応 unit test を test_dag_invariants.py に追加する予定。

Refs #9
