--- backend/tests/domain/test_workflow.py
@@ -1,1083 +1,0 @@
-"""Unit + integration tests for the Workflow aggregate root.
-
-Covers TC-UT-WF-001〜060 and TC-IT-WF-001〜003 from
-``docs/features/workflow/test-design.md``. Tests are grouped into ``Test*``
-classes by feature surface (construction, name, DAG invariants, capacity,
-mutators, from_dict, frozen contract, NotifyChannel SSRF G1〜G10, masking,
-helper independence, lifecycle integration). Each test docstring carries the
-trace anchor (TC-ID, REQ-ID, MSG-ID where applicable).
-
-Integration scenarios live in this file rather than under ``integration/``
-because the aggregate is pure domain (zero external I/O) — the test-design
-intentionally consolidates "Aggregate-internal round-trip" cases here.
-"""
-
-from __future__ import annotations
-
-import json
-from typing import cast
-from uuid import uuid4
-
-import pytest
-from bakufu.domain import workflow as _workflow_module
-from bakufu.domain.exceptions import (
-    StageInvariantViolation,
-    WorkflowInvariantViolation,
-)
-from bakufu.domain.value_objects import (
-    NotifyChannel,
-    Role,
-    StageKind,
-    TransitionCondition,
-    mask_discord_webhook,
-)
-from bakufu.domain.workflow import MAX_STAGES, MAX_TRANSITIONS, Stage, Workflow
-from pydantic import ValidationError
-
-from tests.factories.workflow import (
-    DEFAULT_DISCORD_WEBHOOK,
-    build_v_model_payload,
-    make_notify_channel,
-    make_stage,
-    make_transition,
-    make_workflow,
-)
-
-# Workflow's module-level invariant helpers are private by Python convention
-# (leading underscore) but are intentionally imported by tests under the
-# Confirmation F twin-defense contract. Indirect attribute access bypasses
-# pyright's ``reportPrivateUsage`` warning for this design-mandated path while
-# preserving runtime behavior identical to a direct import.
-_validate_capacity = _workflow_module._validate_capacity  # pyright: ignore[reportPrivateUsage]
-_validate_dag_reachability = _workflow_module._validate_dag_reachability  # pyright: ignore[reportPrivateUsage]
-_validate_dag_sink_exists = _workflow_module._validate_dag_sink_exists  # pyright: ignore[reportPrivateUsage]
-_validate_entry_in_stages = _workflow_module._validate_entry_in_stages  # pyright: ignore[reportPrivateUsage]
-_validate_external_review_notify = _workflow_module._validate_external_review_notify  # pyright: ignore[reportPrivateUsage]
-_validate_required_role_non_empty = _workflow_module._validate_required_role_non_empty  # pyright: ignore[reportPrivateUsage]
-_validate_stage_id_unique = _workflow_module._validate_stage_id_unique  # pyright: ignore[reportPrivateUsage]
-_validate_transition_determinism = _workflow_module._validate_transition_determinism  # pyright: ignore[reportPrivateUsage]
-_validate_transition_refs = _workflow_module._validate_transition_refs  # pyright: ignore[reportPrivateUsage]
-
-# ===========================================================================
-# REQ-WF-001 — Construction & name normalization
-# ===========================================================================
-
-
-class TestWorkflowConstruction:
-    """Minimal Workflow contract (TC-UT-WF-001)."""
-
-    def test_single_stage_workflow_constructs(self) -> None:
-        """TC-UT-WF-001: 1-stage workflow with entry == sink succeeds."""
-        wf = make_workflow()
-        assert len(wf.stages) == 1 and len(wf.transitions) == 0
-
-    def test_entry_stage_id_resolves_to_first_stage(self) -> None:
-        """TC-UT-WF-001: factory's default entry_stage_id matches the lone Stage id."""
-        wf = make_workflow()
-        assert wf.entry_stage_id == wf.stages[0].id
-
-
-class TestWorkflowName:
-    """Workflow.name length and normalization (TC-UT-WF-011 / 012 / 037)."""
-
-    @pytest.mark.parametrize("valid_length", [1, 80])
-    def test_accepts_lower_and_upper_boundary(self, valid_length: int) -> None:
-        """TC-UT-WF-011: 1-char and 80-char names construct."""
-        wf = make_workflow(name="a" * valid_length)
-        assert len(wf.name) == valid_length
-
-    @pytest.mark.parametrize("invalid_name", ["", "a" * 81, "   "])
-    def test_rejects_zero_eightyone_or_whitespace_only(self, invalid_name: str) -> None:
-        """TC-UT-WF-011: 0/81/whitespace-only names raise WorkflowInvariantViolation."""
-        with pytest.raises(WorkflowInvariantViolation) as excinfo:
-            make_workflow(name=invalid_name)
-        assert excinfo.value.kind == "name_range"
-
-    def test_strips_surrounding_whitespace(self) -> None:
-        """TC-UT-WF-012: leading/trailing whitespace is stripped (NFC pipeline)."""
-        wf = make_workflow(name="  V モデル開発フロー  ")
-        assert wf.name == "V モデル開発フロー"
-
-    def test_msg_wf_001_for_oversized_name_matches_exact_wording(self) -> None:
-        """TC-UT-WF-037: MSG-WF-001 wording matches '[FAIL] Workflow name ...'."""
-        with pytest.raises(WorkflowInvariantViolation) as excinfo:
-            make_workflow(name="a" * 81)
-        assert excinfo.value.message == "[FAIL] Workflow name must be 1-80 characters (got 81)"
-
-
-# ===========================================================================
-# REQ-WF-005 — DAG invariants 7 種 (assertions on aggregate construction)
-# ===========================================================================
-
-
-class TestEntryStageId:
-    """REQ-WF-005-① / TC-UT-WF-002 / 038."""
-
-    def test_unknown_entry_raises_entry_not_in_stages(self) -> None:
-        """TC-UT-WF-002: entry_stage_id outside stages raises entry_not_in_stages."""
-        stage = make_stage()
-        with pytest.raises(WorkflowInvariantViolation) as excinfo:
-            make_workflow(stages=[stage], entry_stage_id=uuid4())
-        assert excinfo.value.kind == "entry_not_in_stages"
-
-    def test_msg_wf_002_includes_entry_stage_id(self) -> None:
-        """TC-UT-WF-038: MSG-WF-002 wording carries the offending entry_stage_id."""
-        stage = make_stage()
-        unknown = uuid4()
-        with pytest.raises(WorkflowInvariantViolation) as excinfo:
-            make_workflow(stages=[stage], entry_stage_id=unknown)
-        assert excinfo.value.message == f"[FAIL] entry_stage_id {unknown} not found in stages"
-
-
-class TestUnreachableStage:
-    """REQ-WF-005-④ / TC-UT-WF-003 / 039 — BFS reachability."""
-
-    def test_orphan_stage_raises_unreachable_stage(self) -> None:
-        """TC-UT-WF-003: stage not reachable from entry raises unreachable_stage."""
-        s0 = make_stage()
-        s1 = make_stage()
-        s2 = make_stage()  # orphan, no edges in
-        edge = make_transition(from_stage_id=s0.id, to_stage_id=s1.id)
-        with pytest.raises(WorkflowInvariantViolation) as excinfo:
-            make_workflow(
-                stages=[s0, s1, s2],
-                transitions=[edge],
-                entry_stage_id=s0.id,
-            )
-        assert excinfo.value.kind == "unreachable_stage"
-
-    def test_msg_wf_003_lists_unreachable_stage_ids(self) -> None:
-        """TC-UT-WF-039: MSG-WF-003 includes the unreachable stage's id."""
-        s0 = make_stage()
-        s1 = make_stage()
-        orphan = make_stage()
-        edge = make_transition(from_stage_id=s0.id, to_stage_id=s1.id)
-        with pytest.raises(WorkflowInvariantViolation) as excinfo:
-            make_workflow(stages=[s0, s1, orphan], transitions=[edge], entry_stage_id=s0.id)
-        assert str(orphan.id) in excinfo.value.message
-
-
-class TestSinkStage:
-    """REQ-WF-005-⑤ / TC-UT-WF-004 / 040 — at least one sink Stage."""
-
-    def test_pure_cycle_raises_no_sink_stage(self) -> None:
-        """TC-UT-WF-004: 3 stages all with outgoing edges (cycle) raises no_sink_stage."""
-        s0 = make_stage()
-        s1 = make_stage()
-        s2 = make_stage()
-        e1 = make_transition(from_stage_id=s0.id, to_stage_id=s1.id)
-        e2 = make_transition(from_stage_id=s1.id, to_stage_id=s2.id)
-        e3 = make_transition(from_stage_id=s2.id, to_stage_id=s0.id)
-        with pytest.raises(WorkflowInvariantViolation) as excinfo:
-            make_workflow(
-                stages=[s0, s1, s2],
-                transitions=[e1, e2, e3],
-                entry_stage_id=s0.id,
-            )
-        assert excinfo.value.kind == "no_sink_stage"
-
-    def test_msg_wf_004_starts_with_no_sink_prefix(self) -> None:
-        """TC-UT-WF-040: MSG-WF-004 wording starts with '[FAIL] No sink stage'."""
-        s0 = make_stage()
-        s1 = make_stage()
-        e1 = make_transition(from_stage_id=s0.id, to_stage_id=s1.id)
-        e2 = make_transition(from_stage_id=s1.id, to_stage_id=s0.id)
-        with pytest.raises(WorkflowInvariantViolation) as excinfo:
-            make_workflow(stages=[s0, s1], transitions=[e1, e2], entry_stage_id=s0.id)
-        assert excinfo.value.message.startswith("[FAIL] No sink stage")
-
-
-class TestTransitionDeterminism:
-    """REQ-WF-005-③ / TC-UT-WF-005 / 041 — (from, condition) uniqueness."""
-
-    def test_duplicate_from_condition_raises_transition_duplicate(self) -> None:
-        """TC-UT-WF-005: same (from_stage_id, APPROVED) twice raises transition_duplicate."""
-        s0 = make_stage()
-        s1 = make_stage()
-        s2 = make_stage()
-        e1 = make_transition(
-            from_stage_id=s0.id, to_stage_id=s1.id, condition=TransitionCondition.APPROVED
-        )
-        e2 = make_transition(
-            from_stage_id=s0.id, to_stage_id=s2.id, condition=TransitionCondition.APPROVED
-        )
-        # s0 has 2 outgoing → s2 has none from chain. Add a forward edge to make
-        # a sink; but the duplicate check should fire before sink check.
-        with pytest.raises(WorkflowInvariantViolation) as excinfo:
-            make_workflow(
-                stages=[s0, s1, s2],
-                transitions=[e1, e2],
-                entry_stage_id=s0.id,
-            )
-        assert excinfo.value.kind == "transition_duplicate"
-
-    def test_msg_wf_005_includes_from_and_condition(self) -> None:
-        """TC-UT-WF-041: MSG-WF-005 wording carries from_stage and condition."""
-        s0 = make_stage()
-        s1 = make_stage()
-        s2 = make_stage()
-        e1 = make_transition(
-            from_stage_id=s0.id, to_stage_id=s1.id, condition=TransitionCondition.APPROVED
-        )
-        e2 = make_transition(
-            from_stage_id=s0.id, to_stage_id=s2.id, condition=TransitionCondition.APPROVED
-        )
-        with pytest.raises(WorkflowInvariantViolation) as excinfo:
-            make_workflow(stages=[s0, s1, s2], transitions=[e1, e2], entry_stage_id=s0.id)
-        assert "APPROVED" in excinfo.value.message
-        assert str(s0.id) in excinfo.value.message
-
-
-class TestExternalReviewNotify:
-    """REQ-WF-007-② / TC-UT-WF-006a / 006b / 042 — twin-defense."""
-
-    def test_006a_stage_self_path_rejects_empty_notify(self) -> None:
-        """TC-UT-WF-006a: Stage self-validator rejects empty notify_channels.
-
-        Construction with ``kind=EXTERNAL_REVIEW`` and ``notify_channels=[]``
-        raises :class:`StageInvariantViolation` directly from the Stage's
-        own ``model_validator(mode='after')``, before any aggregate check
-        runs — fulfilling the twin-defense Stage-side path.
-        """
-        with pytest.raises(StageInvariantViolation) as excinfo:
-            make_stage(kind=StageKind.EXTERNAL_REVIEW, notify_channels=[])
-        assert excinfo.value.kind == "missing_notify"
-
-    def test_006b_aggregate_helper_rejects_empty_notify_via_direct_call(self) -> None:
-        """TC-UT-WF-006b: ``_validate_external_review_notify`` raises aggregate violation directly.
-
-        Builds an EXTERNAL_REVIEW Stage *with* notify_channels (so its own
-        validator passes), then mutates the resulting list snapshot to empty
-        and calls the aggregate helper to prove it independently catches the
-        violation. This proves the dual code paths (Stage self vs aggregate
-        helper) do not share code.
-        """
-        good = make_stage(kind=StageKind.EXTERNAL_REVIEW)
-        # Reconstruct the same Stage but with notify_channels=[] by going
-        # through ``model_construct`` (skips validators) so the helper sees a
-        # state the Stage validator would have rejected. This is the only way
-        # to reach the aggregate helper independently.
-        stage_without_notify = Stage.model_construct(
-            id=good.id,
-            name=good.name,
-            kind=StageKind.EXTERNAL_REVIEW,
-            required_role=good.required_role,
-            deliverable_template=good.deliverable_template,
-            completion_policy=good.completion_policy,
-            notify_channels=[],
-        )
-        with pytest.raises(WorkflowInvariantViolation) as excinfo:
-            _validate_external_review_notify([stage_without_notify])
-        assert excinfo.value.kind == "missing_notify_aggregate"
-
-    def test_msg_wf_006_includes_stage_id(self) -> None:
-        """TC-UT-WF-042: MSG-WF-006 wording carries the offending stage_id."""
-        with pytest.raises(StageInvariantViolation) as excinfo:
-            make_stage(kind=StageKind.EXTERNAL_REVIEW, notify_channels=[])
-        # Accept either canonical wording variant; we just verify the stage_id
-        # is present so debugging is possible.
-        assert "EXTERNAL_REVIEW stage" in excinfo.value.message
-        assert "must have at least one notify_channel" in excinfo.value.message
-
-
-class TestRequiredRoleNonEmpty:
-    """REQ-WF-007-① / TC-UT-WF-007 / 043 — required_role non-empty."""
-
-    def test_stage_self_path_rejects_empty_required_role(self) -> None:
-        """TC-UT-WF-007: Stage(required_role=frozenset()) raises empty_required_role."""
-        with pytest.raises(StageInvariantViolation) as excinfo:
-            make_stage(required_role=frozenset())
-        assert excinfo.value.kind == "empty_required_role"
-
-    def test_msg_wf_007_includes_stage_id(self) -> None:
-        """TC-UT-WF-043: MSG-WF-007 wording carries the offending stage_id."""
-        with pytest.raises(StageInvariantViolation) as excinfo:
-            make_stage(required_role=frozenset())
-        assert "required_role must not be empty" in excinfo.value.message
-
-    def test_aggregate_helper_rejects_empty_required_role(self) -> None:
-        """`_validate_required_role_non_empty` rejects bypassed stage with empty role set."""
-        good = make_stage()
-        bad = Stage.model_construct(
-            id=good.id,
-            name=good.name,
-            kind=StageKind.WORK,
-            required_role=frozenset(),
-            deliverable_template=good.deliverable_template,
-            completion_policy=good.completion_policy,
-            notify_channels=[],
-        )
-        with pytest.raises(WorkflowInvariantViolation) as excinfo:
-            _validate_required_role_non_empty([bad])
-        assert excinfo.value.kind == "empty_required_role_aggregate"
-
-
-class TestTransitionRefIntegrity:
-    """REQ-WF-005-② / TC-UT-WF-022 / 045 — Transition refs must point at known Stages."""
-
-    def test_transition_to_unknown_stage_raises(self) -> None:
-        """TC-UT-WF-022: Transition.to_stage_id outside stages raises transition_ref_invalid."""
-        s0 = make_stage()
-        bad_to = uuid4()
-        edge = make_transition(from_stage_id=s0.id, to_stage_id=bad_to)
-        with pytest.raises(WorkflowInvariantViolation) as excinfo:
-            make_workflow(stages=[s0], transitions=[edge], entry_stage_id=s0.id)
-        assert excinfo.value.kind == "transition_ref_invalid"
-
-    def test_msg_wf_009_includes_from_and_to(self) -> None:
-        """TC-UT-WF-045: MSG-WF-009 wording carries from / to stage ids."""
-        s0 = make_stage()
-        bad_to = uuid4()
-        edge = make_transition(from_stage_id=s0.id, to_stage_id=bad_to)
-        with pytest.raises(WorkflowInvariantViolation) as excinfo:
-            make_workflow(stages=[s0], transitions=[edge], entry_stage_id=s0.id)
-        assert "Transition references unknown stage" in excinfo.value.message
-        assert str(bad_to) in excinfo.value.message
-
-
-class TestBFSCycleSafety:
-    """TC-UT-WF-023 — BFS terminates on cyclic graphs without infinite loop."""
-
-    def test_helper_terminates_on_cycle(self) -> None:
-        """`_validate_dag_reachability` returns without hanging on a 3-stage cycle."""
-        s0 = make_stage()
-        s1 = make_stage()
-        s2 = make_stage()
-        e1 = make_transition(from_stage_id=s0.id, to_stage_id=s1.id)
-        e2 = make_transition(from_stage_id=s1.id, to_stage_id=s2.id)
-        e3 = make_transition(from_stage_id=s2.id, to_stage_id=s0.id)
-        # All stages reachable; reachability check should not raise.
-        _validate_dag_reachability([s0, s1, s2], [e1, e2, e3], s0.id)
-
-
-# ===========================================================================
-# Mutators (REQ-WF-002 / 003 / 004) + capacity
-# ===========================================================================
-
-
-class TestAddStage:
-    """REQ-WF-002 / TC-UT-WF-014 / 044.
-
-    Note on TC-UT-WF-013: ``add_stage`` in isolation cannot succeed because a
-    newly-added Stage with no incoming Transition is orphaned and the
-    aggregate-level ``_validate_dag_reachability`` rejects it. The "appends"
-    success path is exercised through ``from_dict`` (TC-IT-WF-001) and the
-    payload-construction tests in ``TestWorkflowLifecycleIntegration`` —
-    granular mutators are designed for **constrained** changes. We retain the
-    failure-path tests here to lock the rebuild contract.
-    """
-
-    def test_duplicate_stage_id_raises_stage_duplicate(self) -> None:
-        """TC-UT-WF-014: add_stage with existing id raises stage_duplicate."""
-        wf = make_workflow()
-        existing = wf.stages[0]
-        duplicate = make_stage(stage_id=existing.id)
-        with pytest.raises(WorkflowInvariantViolation) as excinfo:
-            wf.add_stage(duplicate)
-        assert excinfo.value.kind == "stage_duplicate"
-
-    def test_msg_wf_008_for_duplicate_includes_stage_id(self) -> None:
-        """TC-UT-WF-044: MSG-WF-008 wording carries the duplicate stage_id."""
-        wf = make_workflow()
-        existing = wf.stages[0]
-        duplicate = make_stage(stage_id=existing.id)
-        with pytest.raises(WorkflowInvariantViolation) as excinfo:
-            wf.add_stage(duplicate)
-        assert excinfo.value.message == f"[FAIL] Stage id duplicate: {existing.id}"
-
-
-class TestAddStageCapacity:
-    """REQ-WF-002 capacity / TC-UT-WF-015."""
-
-    def test_overflow_raises_capacity_exceeded(self) -> None:
-        """TC-UT-WF-015: bulk-importing >MAX_STAGES via from_dict raises capacity_exceeded.
-
-        We can't reach 30 via add_stage easily because each addition must keep
-        all stages reachable; the boundary is best demonstrated through a
-        bulk-import payload.
-        """
-        stages: list[dict[str, object]] = []
-        prev_id: str | None = None
-        first_id: str | None = None
-        for _ in range(MAX_STAGES + 1):
-            sid = str(uuid4())
-            stages.append(
-                {
-                    "id": sid,
-                    "name": "S",
-                    "kind": StageKind.WORK.value,
-                    "required_role": [Role.DEVELOPER.value],
-                    "deliverable_template": "",
-                    "completion_policy": {"kind": "manual", "description": ""},
-                    "notify_channels": [],
-                }
-            )
-            if first_id is None:
-                first_id = sid
-            prev_id = sid
-        # Build forward APPROVED chain so we have at least valid topology.
-        transitions: list[dict[str, object]] = []
-        previous: str | None = None
-        for stage in stages:
-            current = stage["id"]
-            if previous is not None:
-                transitions.append(
-                    {
-                        "id": str(uuid4()),
-                        "from_stage_id": previous,
-                        "to_stage_id": current,
-                        "condition": TransitionCondition.APPROVED.value,
-                        "label": "",
-                    }
-                )
-            previous = str(current)
-        _ = prev_id  # unused: linter satisfaction
-        payload = {
-            "id": str(uuid4()),
-            "name": "overflow",
-            "stages": stages,
-            "transitions": transitions,
-            "entry_stage_id": first_id,
-        }
-        with pytest.raises(WorkflowInvariantViolation) as excinfo:
-            Workflow.from_dict(payload)
-        assert excinfo.value.kind == "capacity_exceeded"
-
-
-class TestAddTransition:
-    """REQ-WF-003 / TC-UT-WF-016 / 017."""
-
-    def test_appends_transition_to_list(self) -> None:
-        """TC-UT-WF-016: add_transition returns a new Workflow with the edge appended.
-
-        Pre-state: 3-stage chain s0→s1→s2 with two forward APPROVED edges. s2
-        is the sink. Add a REJECTED back-edge s1→s0 — s2 stays the sole sink,
-        the new (s1, REJECTED) pair is unique, all stages remain reachable.
-        """
-        s0 = make_stage()
-        s1 = make_stage()
-        s2 = make_stage()
-        e0 = make_transition(
-            from_stage_id=s0.id, to_stage_id=s1.id, condition=TransitionCondition.APPROVED
-        )
-        e1 = make_transition(
-            from_stage_id=s1.id, to_stage_id=s2.id, condition=TransitionCondition.APPROVED
-        )
-        wf = make_workflow(stages=[s0, s1, s2], transitions=[e0, e1], entry_stage_id=s0.id)
-        e_back = make_transition(
-            from_stage_id=s1.id, to_stage_id=s0.id, condition=TransitionCondition.REJECTED
-        )
-        updated = wf.add_transition(e_back)
-        assert len(updated.transitions) == 3
-
-    def test_does_not_mutate_original(self) -> None:
-        """TC-UT-WF-016: caller's Workflow stays at 2 transitions after add path."""
-        s0 = make_stage()
-        s1 = make_stage()
-        s2 = make_stage()
-        e0 = make_transition(from_stage_id=s0.id, to_stage_id=s1.id)
-        e1 = make_transition(from_stage_id=s1.id, to_stage_id=s2.id)
-        wf = make_workflow(stages=[s0, s1, s2], transitions=[e0, e1], entry_stage_id=s0.id)
-        e_back = make_transition(
-            from_stage_id=s1.id, to_stage_id=s0.id, condition=TransitionCondition.REJECTED
-        )
-        wf.add_transition(e_back)
-        assert len(wf.transitions) == 2
-
-    def test_dangling_ref_raises_transition_ref_invalid(self) -> None:
-        """TC-UT-WF-017: add_transition with unknown from/to raises transition_ref_invalid."""
-        wf = make_workflow()
-        bad_edge = make_transition(from_stage_id=wf.stages[0].id, to_stage_id=uuid4())
-        with pytest.raises(WorkflowInvariantViolation) as excinfo:
-            wf.add_transition(bad_edge)
-        assert excinfo.value.kind == "transition_ref_invalid"
-
-
-class TestAddTransitionCapacity:
-    """REQ-WF-003 capacity / TC-UT-WF-018."""
-
-    def test_overflow_raises_capacity_exceeded(self) -> None:
-        """TC-UT-WF-018: building >MAX_TRANSITIONS via from_dict raises capacity_exceeded."""
-        # 2 stages with MAX_TRANSITIONS+1 edges between them, varying conditions.
-        # Since (from, condition) must be unique, we use distinct from/to pairs.
-        stage_ids = [str(uuid4()) for _ in range(2)]
-        stages_payload: list[dict[str, object]] = [
-            {
-                "id": sid,
-                "name": "S",
-                "kind": StageKind.WORK.value,
-                "required_role": [Role.DEVELOPER.value],
-                "deliverable_template": "",
-                "completion_policy": {"kind": "manual", "description": ""},
-                "notify_channels": [],
-            }
-            for sid in stage_ids
-        ]
-        transitions_payload: list[dict[str, object]] = []
-        # Forward APPROVED edge so topology is valid.
-        transitions_payload.append(
-            {
-                "id": str(uuid4()),
-                "from_stage_id": stage_ids[0],
-                "to_stage_id": stage_ids[1],
-                "condition": TransitionCondition.APPROVED.value,
-                "label": "",
-            }
-        )
-        # Padding edges with same (from, to) but different conditions would
-        # collide on (from, condition) uniqueness once we exhaust them. With
-        # only 4 conditions we can't reach 60+ uniquely. Easier: alternate
-        # from/to pairs by adding self-loop-style edges (allowed at the
-        # transition level, only ref integrity matters) with each rotation.
-        # We only need to trip capacity, not pass validation, so use distinct
-        # ids on dangling refs — but that fails on transition_ref_invalid.
-        # Instead, fabricate transitions referencing the same valid (s0→s1)
-        # pair across all 4 conditions, and exceed by repeating until count > MAX.
-        for _ in range(MAX_TRANSITIONS):
-            transitions_payload.append(
-                {
-                    "id": str(uuid4()),
-                    "from_stage_id": stage_ids[0],
-                    "to_stage_id": stage_ids[1],
-                    "condition": TransitionCondition.APPROVED.value,
-                    "label": "",
-                }
-            )
-        payload = {
-            "id": str(uuid4()),
-            "name": "overflow",
-            "stages": stages_payload,
-            "transitions": transitions_payload,
-            "entry_stage_id": stage_ids[0],
-        }
-        with pytest.raises(WorkflowInvariantViolation) as excinfo:
-            Workflow.from_dict(payload)
-        # Could be capacity_exceeded (61 transitions) or transition_duplicate
-        # (multiple APPROVED from same from). Capacity is checked first.
-        assert excinfo.value.kind == "capacity_exceeded"
-
-
-class TestRemoveStage:
-    """REQ-WF-004 / TC-UT-WF-009 / 019 / 020 / 021 / 046."""
-
-    def test_remove_entry_stage_raises_cannot_remove_entry(self) -> None:
-        """TC-UT-WF-009: remove_stage(entry_stage_id) raises cannot_remove_entry."""
-        wf = make_workflow()
-        with pytest.raises(WorkflowInvariantViolation) as excinfo:
-            wf.remove_stage(wf.entry_stage_id)
-        assert excinfo.value.kind == "cannot_remove_entry"
-
-    def test_msg_wf_010_includes_stage_id(self) -> None:
-        """TC-UT-WF-046: MSG-WF-010 wording carries the entry stage_id."""
-        wf = make_workflow()
-        with pytest.raises(WorkflowInvariantViolation) as excinfo:
-            wf.remove_stage(wf.entry_stage_id)
-        assert excinfo.value.message == f"[FAIL] Cannot remove entry stage: {wf.entry_stage_id}"
-
-    def test_unknown_stage_id_raises_stage_not_found(self) -> None:
-        """TC-UT-WF-019: remove_stage(unknown) raises stage_not_found with MSG-WF-012."""
-        wf = make_workflow()
-        unknown = uuid4()
-        with pytest.raises(WorkflowInvariantViolation) as excinfo:
-            wf.remove_stage(unknown)
-        assert excinfo.value.kind == "stage_not_found"
-        assert excinfo.value.message == f"[FAIL] Stage not found in workflow: stage_id={unknown}"
-
-    def test_cascades_incident_transitions(self) -> None:
-        """TC-UT-WF-020: removing a Stage also drops Transitions referencing it."""
-        s0 = make_stage()
-        s1 = make_stage()
-        s2 = make_stage()
-        e0 = make_transition(from_stage_id=s0.id, to_stage_id=s1.id)
-        e1 = make_transition(from_stage_id=s1.id, to_stage_id=s2.id)
-        wf = make_workflow(stages=[s0, s1, s2], transitions=[e0, e1], entry_stage_id=s0.id)
-        # Removing the trailing stage s2 leaves s0→s1 valid (s1 becomes sink).
-        updated = wf.remove_stage(s2.id)
-        assert len(updated.stages) == 2 and len(updated.transitions) == 1
-
-
-class TestPreValidateRollback:
-    """Confirmation A — failed mutators leave original Workflow unchanged."""
-
-    def test_failed_add_stage_keeps_original(self) -> None:
-        """TC-UT-WF-008: failed add_stage does not mutate caller's Workflow."""
-        wf = make_workflow()
-        existing = wf.stages[0]
-        with pytest.raises(WorkflowInvariantViolation):
-            wf.add_stage(make_stage(stage_id=existing.id))
-        assert len(wf.stages) == 1
-
-    def test_failed_add_transition_keeps_original(self) -> None:
-        """TC-UT-WF-030: failed add_transition does not mutate caller's Workflow."""
-        wf = make_workflow()
-        bad = make_transition(from_stage_id=wf.stages[0].id, to_stage_id=uuid4())
-        with pytest.raises(WorkflowInvariantViolation):
-            wf.add_transition(bad)
-        assert wf.transitions == []
-
-    def test_failed_remove_stage_keeps_original(self) -> None:
-        """TC-UT-WF-031: failed remove_stage does not mutate caller's Workflow."""
-        wf = make_workflow()
-        with pytest.raises(WorkflowInvariantViolation):
-            wf.remove_stage(uuid4())
-        assert len(wf.stages) == 1
-
-
-# ===========================================================================
-# from_dict bulk-import (REQ-WF-006 / T1 attack-surface)
-# ===========================================================================
-
-
-class TestFromDict:
-    """REQ-WF-006 / TC-UT-WF-024〜027 / 047."""
-
-    def test_unknown_role_in_payload_raises(self) -> None:
-        """TC-UT-WF-024: from_dict with role='UNKNOWN_ROLE' raises a violation."""
-        payload = build_v_model_payload()
-        stages = cast("list[dict[str, object]]", payload["stages"])
-        stages[0]["required_role"] = ["UNKNOWN_ROLE"]
-        with pytest.raises((ValidationError, WorkflowInvariantViolation)):
-            Workflow.from_dict(payload)
-
-    def test_invalid_uuid_raises(self) -> None:
-        """TC-UT-WF-025: from_dict with id='not-a-uuid' raises ValidationError."""
-        payload = build_v_model_payload()
-        payload["id"] = "not-a-uuid"
-        with pytest.raises((ValidationError, WorkflowInvariantViolation)):
-            Workflow.from_dict(payload)
-
-    def test_missing_entry_stage_id_raises(self) -> None:
-        """TC-UT-WF-026: from_dict without entry_stage_id raises ValidationError."""
-        payload = build_v_model_payload()
-        del payload["entry_stage_id"]
-        with pytest.raises((ValidationError, WorkflowInvariantViolation)):
-            Workflow.from_dict(payload)
-
-    def test_stage_index_appears_in_detail_on_stage_violation(self) -> None:
-        """TC-UT-WF-027: from_dict failure surfaces the offending stage index."""
-        payload = build_v_model_payload()
-        stages = cast("list[dict[str, object]]", payload["stages"])
-        # Empty required_role on stage index 2 (the third Stage).
-        stages[2]["required_role"] = []
-        with pytest.raises(WorkflowInvariantViolation) as excinfo:
-            Workflow.from_dict(payload)
-        assert excinfo.value.kind == "from_dict_invalid"
-        assert excinfo.value.detail.get("stage_index") == 2
-
-    def test_msg_wf_011_starts_with_payload_invalid(self) -> None:
-        """TC-UT-WF-047: MSG-WF-011 starts with '[FAIL] from_dict payload invalid:'."""
-        with pytest.raises(WorkflowInvariantViolation) as excinfo:
-            Workflow.from_dict("not-a-dict")
-        assert excinfo.value.message.startswith("[FAIL] from_dict payload invalid:")
-
-
-# ===========================================================================
-# Frozen / extra='forbid'
-# ===========================================================================
-
-
-class TestFrozenContract:
-    """TC-UT-WF-032 — Workflow / Stage / Transition / NotifyChannel frozen."""
-
-    def test_workflow_rejects_attribute_assignment(self) -> None:
-        wf = make_workflow()
-        with pytest.raises(ValidationError):
-            wf.name = "改竄"  # type: ignore[misc]
-
-    def test_stage_rejects_attribute_assignment(self) -> None:
-        stage = make_stage()
-        with pytest.raises(ValidationError):
-            stage.kind = StageKind.EXTERNAL_REVIEW  # type: ignore[misc]
-
-    def test_transition_rejects_attribute_assignment(self) -> None:
-        s0 = make_stage()
-        s1 = make_stage()
-        edge = make_transition(from_stage_id=s0.id, to_stage_id=s1.id)
-        with pytest.raises(ValidationError):
-            edge.condition = TransitionCondition.REJECTED  # type: ignore[misc]
-
-    def test_notify_channel_rejects_attribute_assignment(self) -> None:
-        channel = make_notify_channel()
-        with pytest.raises(ValidationError):
-            channel.target = "https://discord.com/api/webhooks/2/3"  # type: ignore[misc]
-
-
-class TestExtraForbid:
-    """TC-UT-WF-033 — extra='forbid' rejects unknown fields."""
-
-    def test_workflow_rejects_unknown_field(self) -> None:
-        payload = build_v_model_payload()
-        payload["unknown_field"] = "x"
-        with pytest.raises((ValidationError, WorkflowInvariantViolation)):
-            Workflow.from_dict(payload)
-
-
-# ===========================================================================
-# NotifyChannel SSRF G1〜G10 (Confirmation G)
-# ===========================================================================
-
-
-class TestNotifyChannelSSRF:
-    """TC-UT-WF-034〜036, 048〜054 — full G1〜G10 rejection coverage."""
-
-    @pytest.mark.parametrize(
-        "bad_target",
-        [
-            # G3: HTTPS強制
-            "http://discord.com/api/webhooks/123/abc-DEF_xyz",
-        ],
-    )
-    def test_034_https_only(self, bad_target: str) -> None:
-        """TC-UT-WF-034 / G3: scheme must be 'https'."""
-        with pytest.raises(ValidationError):
-            NotifyChannel(kind="discord", target=bad_target)
-
-    @pytest.mark.parametrize(
-        "bad_target",
-        [
-            "https://discord.com.evil.example/api/webhooks/123/abc",
-            "https://evil-discord.com/api/webhooks/123/abc",
-            "https://api.discord.com/api/webhooks/123/abc",
-        ],
-    )
-    def test_035_hostname_exact_match(self, bad_target: str) -> None:
-        """TC-UT-WF-035 / G4: hostname must equal 'discord.com' exactly."""
-        with pytest.raises(ValidationError):
-            NotifyChannel(kind="discord", target=bad_target)
-
-    @pytest.mark.parametrize(
-        "bad_target",
-        [
-            "https://discord.com/",
-            "https://discord.com/api/webhooks/",
-        ],
-    )
-    def test_036_path_must_be_present(self, bad_target: str) -> None:
-        """TC-UT-WF-036 / G7: path must match /api/webhooks/<id>/<token>."""
-        with pytest.raises(ValidationError):
-            NotifyChannel(kind="discord", target=bad_target)
-
-    def test_048_token_at_g7_cap_succeeds_overflow_rejected(self) -> None:
-        """TC-UT-WF-048 / G1+G7: 100-char token (G7 cap) works, 101+ rejected.
-
-        Realistic upper bound on a *valid* Discord webhook URL is reached via
-        G7 (token ≤ 100 chars). We verify that the maximum-permitted shape
-        constructs successfully and that any overflow — which simultaneously
-        trips G1 (>500 chars) for sufficiently long tokens or G7 alone —
-        produces a ``ValidationError``.
-        """
-        base = "https://discord.com/api/webhooks/123456789/"
-        valid = base + "a" * 100  # token at G7 cap, total well under G1's 500.
-        channel = NotifyChannel(kind="discord", target=valid)
-        assert channel.target == valid
-        # 101-char token violates G7. Total length still well under 500.
-        with pytest.raises(ValidationError):
-            NotifyChannel(kind="discord", target=base + "a" * 101)
-        # Pad the URL beyond 500 chars by generating a giant *non-discord* URL
-        # so we can prove G1 fires independently when G7's token rule is also
-        # violated.
-        oversized = "https://discord.com/api/webhooks/1/" + "a" * 500
-        assert len(oversized) > 500
-        with pytest.raises(ValidationError):
-            NotifyChannel(kind="discord", target=oversized)
-
-    @pytest.mark.parametrize(
-        "bad_target",
-        [
-            "https://discord.com:80/api/webhooks/123/abc-DEF_xyz",
-            "https://discord.com:8443/api/webhooks/123/abc-DEF_xyz",
-            "https://discord.com:8080/api/webhooks/123/abc-DEF_xyz",
-        ],
-    )
-    def test_049_port_must_be_none_or_443(self, bad_target: str) -> None:
-        """TC-UT-WF-049 / G5: port restricted to {None, 443}."""
-        with pytest.raises(ValidationError):
-            NotifyChannel(kind="discord", target=bad_target)
-
-    @pytest.mark.parametrize(
-        "bad_target",
-        [
-            "https://attacker@discord.com/api/webhooks/123/abc-DEF_xyz",
-            "https://user:pass@discord.com/api/webhooks/123/abc-DEF_xyz",
-        ],
-    )
-    def test_050_userinfo_rejected(self, bad_target: str) -> None:
-        """TC-UT-WF-050 / G6: userinfo (user/password) rejected."""
-        with pytest.raises(ValidationError):
-            NotifyChannel(kind="discord", target=bad_target)
-
-    @pytest.mark.parametrize(
-        "bad_target",
-        [
-            "https://discord.com/api/webhooks/abc/def",  # id non-numeric
-            "https://discord.com/api/webhooks/123/!@#",  # token bad chars
-            "https://discord.com/api/webhooks/" + ("0" * 31) + "/abc",  # id 31 digits
-            "https://discord.com/api/webhooks/123/" + ("a" * 101),  # token 101 chars
-            "https://discord.com/api/webhooks/123/abc/extra",  # extra path segment
-        ],
-    )
-    def test_051_path_regex_fullmatch(self, bad_target: str) -> None:
-        """TC-UT-WF-051 / G7: path regex must fullmatch /api/webhooks/<id>/<token>."""
-        with pytest.raises(ValidationError):
-            NotifyChannel(kind="discord", target=bad_target)
-
-    def test_052_query_rejected(self) -> None:
-        """TC-UT-WF-052 / G8: query string rejected."""
-        with pytest.raises(ValidationError):
-            NotifyChannel(
-                kind="discord",
-                target="https://discord.com/api/webhooks/123/abc?override=x",
-            )
-
-    def test_053_fragment_rejected(self) -> None:
-        """TC-UT-WF-053 / G9: fragment rejected."""
-        with pytest.raises(ValidationError):
-            NotifyChannel(
-                kind="discord",
-                target="https://discord.com/api/webhooks/123/abc#frag",
-            )
-
-    @pytest.mark.parametrize(
-        "bad_target",
-        [
-            "https://discord.com/API/WEBHOOKS/123/abc",  # uppercase API/WEBHOOKS
-            "https://discord.com/Api/Webhooks/123/abc",  # mixed case
-        ],
-    )
-    def test_054_path_case_sensitive(self, bad_target: str) -> None:
-        """TC-UT-WF-054 / G10: path case-sensitive (only lowercase /api/webhooks/)."""
-        with pytest.raises(ValidationError):
-            NotifyChannel(kind="discord", target=bad_target)
-
-
-class TestNotifyChannelKindMVP:
-    """TC-UT-WF-055 / 056 — MVP only accepts kind='discord'."""
-
-    @pytest.mark.parametrize("bad_kind", ["slack", "email"])
-    def test_non_discord_kind_rejected(self, bad_kind: str) -> None:
-        """TC-UT-WF-055/056: kind='slack' or 'email' raises ValidationError."""
-        with pytest.raises(ValidationError):
-            NotifyChannel.model_validate({"kind": bad_kind, "target": DEFAULT_DISCORD_WEBHOOK})
-
-
-# ===========================================================================
-# NotifyChannel secret masking (Confirmation G "target のシークレット扱い")
-# ===========================================================================
-
-
-class TestNotifyChannelMasking:
-    """TC-UT-WF-057 / 058 / 059 — token masked on JSON serialization & exceptions."""
-
-    def test_057_model_dump_json_mode_masks_token(self) -> None:
-        """TC-UT-WF-057: model_dump(mode='json') replaces token with REDACTED."""
-        channel = make_notify_channel()
-        dumped = channel.model_dump(mode="json")
-        assert "<REDACTED:DISCORD_WEBHOOK>" in dumped["target"]
-        assert "SyntheticToken_-abcXYZ" not in dumped["target"]
-
-    def test_057_model_dump_python_mode_preserves_token(self) -> None:
-        """TC-UT-WF-057: model_dump(mode='python') keeps raw target for in-process use."""
-        channel = make_notify_channel()
-        dumped = channel.model_dump()
-        assert dumped["target"] == DEFAULT_DISCORD_WEBHOOK
-
-    def test_058_model_dump_json_workflow_scans_clean(self) -> None:
-        """TC-UT-WF-058: workflow.model_dump_json() shows no plaintext token segment."""
-        wf_payload = build_v_model_payload()
-        wf = Workflow.from_dict(wf_payload)
-        json_text = wf.model_dump_json()
-        # Token segment "SyntheticToken_-abcXYZ" must not appear in JSON output.
-        assert "SyntheticToken_-abcXYZ" not in json_text
-        assert "<REDACTED:DISCORD_WEBHOOK>" in json_text
-        # Sanity check: the dumped JSON parses back as well-formed JSON.
-        parsed = json.loads(json_text)
-        assert parsed["name"] == "V モデル開発フロー"
-
-    def test_059_exception_detail_does_not_leak_token(self) -> None:
-        """TC-UT-WF-059: WorkflowInvariantViolation message/detail mask Discord tokens."""
-        # Construct an exception manually with a webhook URL embedded in detail
-        # to prove auto-masking on init.
-        fake_message = (
-            f"[FAIL] something happened with https://discord.com/api/webhooks/123/{'a' * 30}"
-        )
-        exc = WorkflowInvariantViolation(
-            kind="from_dict_invalid",
-            message=fake_message,
-            detail={
-                "url": f"https://discord.com/api/webhooks/123/{'b' * 30}",
-                "nested": {"u": f"https://discord.com/api/webhooks/123/{'c' * 30}"},
-            },
-        )
-        assert "a" * 30 not in exc.message
-        # Verify masking applied to nested detail dict too.
-        nested = cast("dict[str, object]", exc.detail["nested"])
-        assert "c" * 30 not in str(nested["u"])
-
-    def test_mask_helper_is_idempotent(self) -> None:
-        """Applying mask_discord_webhook twice yields the same result (smoke)."""
-        original = f"https://discord.com/api/webhooks/123/{'x' * 20}"
-        once = mask_discord_webhook(original)
-        twice = mask_discord_webhook(once)
-        assert once == twice
-
-
-# ===========================================================================
-# Helper independence (Confirmation F)
-# ===========================================================================
-
-
-class TestHelperIndependence:
-    """TC-UT-WF-060 — module-level helpers are import-able and independently raise."""
-
-    def test_capacity_helper_raises_directly(self) -> None:
-        """`_validate_capacity` raises capacity_exceeded when stages > MAX_STAGES."""
-        too_many = [make_stage() for _ in range(MAX_STAGES + 1)]
-        with pytest.raises(WorkflowInvariantViolation) as excinfo:
-            _validate_capacity(too_many, [])
-        assert excinfo.value.kind == "capacity_exceeded"
-
-    def test_stage_id_unique_helper_raises_directly(self) -> None:
-        """`_validate_stage_id_unique` raises stage_duplicate on collisions."""
-        s0 = make_stage()
-        dup = make_stage(stage_id=s0.id)
-        with pytest.raises(WorkflowInvariantViolation) as excinfo:
-            _validate_stage_id_unique([s0, dup])
-        assert excinfo.value.kind == "stage_duplicate"
-
-    def test_entry_in_stages_helper_raises_directly(self) -> None:
-        """`_validate_entry_in_stages` raises when entry not in stages."""
-        s0 = make_stage()
-        with pytest.raises(WorkflowInvariantViolation) as excinfo:
-            _validate_entry_in_stages([s0], uuid4())
-        assert excinfo.value.kind == "entry_not_in_stages"
-
-    def test_transition_refs_helper_raises_directly(self) -> None:
-        """`_validate_transition_refs` raises when from/to point at unknown stages."""
-        s0 = make_stage()
-        bogus = make_transition(from_stage_id=s0.id, to_stage_id=uuid4())
-        with pytest.raises(WorkflowInvariantViolation) as excinfo:
-            _validate_transition_refs([s0], [bogus])
-        assert excinfo.value.kind == "transition_ref_invalid"
-
-    def test_transition_determinism_helper_raises_directly(self) -> None:
-        """`_validate_transition_determinism` raises on (from, condition) collision."""
-        s0 = make_stage()
-        s1 = make_stage()
-        e1 = make_transition(
-            from_stage_id=s0.id, to_stage_id=s1.id, condition=TransitionCondition.APPROVED
-        )
-        e2 = make_transition(
-            from_stage_id=s0.id, to_stage_id=s1.id, condition=TransitionCondition.APPROVED
-        )
-        with pytest.raises(WorkflowInvariantViolation) as excinfo:
-            _validate_transition_determinism([e1, e2])
-        assert excinfo.value.kind == "transition_duplicate"
-
-    def test_dag_sink_helper_raises_directly(self) -> None:
-        """`_validate_dag_sink_exists` raises when no Stage lacks outgoing edges."""
-        s0 = make_stage()
-        s1 = make_stage()
-        e1 = make_transition(from_stage_id=s0.id, to_stage_id=s1.id)
-        e2 = make_transition(from_stage_id=s1.id, to_stage_id=s0.id)
-        with pytest.raises(WorkflowInvariantViolation) as excinfo:
-            _validate_dag_sink_exists([s0, s1], [e1, e2], s0.id)
-        assert excinfo.value.kind == "no_sink_stage"
-
-    def test_dag_reachability_helper_raises_directly(self) -> None:
-        """`_validate_dag_reachability` raises unreachable_stage when orphan exists."""
-        s0 = make_stage()
-        orphan = make_stage()
-        with pytest.raises(WorkflowInvariantViolation) as excinfo:
-            _validate_dag_reachability([s0, orphan], [], s0.id)
-        assert excinfo.value.kind == "unreachable_stage"
-
-
-# ===========================================================================
-# Integration scenarios (TC-IT-WF-001 / 002 / 003)
-# ===========================================================================
-
-
-class TestWorkflowLifecycleIntegration:
-    """Aggregate-internal round-trip + V-model preset + bad-payload variants."""
-
-    def test_v_model_preset_constructs_via_from_dict(self) -> None:
-        """TC-IT-WF-001: 13-stage / 15-transition V-model payload constructs."""
-        wf = Workflow.from_dict(build_v_model_payload())
-        assert len(wf.stages) == 13 and len(wf.transitions) == 15
-
-    def test_v_model_preset_has_external_review_notify_channels(self) -> None:
-        """TC-IT-WF-001: every EXTERNAL_REVIEW Stage has at least one notify_channel."""
-        wf = Workflow.from_dict(build_v_model_payload())
-        review_stages = [s for s in wf.stages if s.kind is StageKind.EXTERNAL_REVIEW]
-        assert all(len(s.notify_channels) >= 1 for s in review_stages)
-
-    def test_round_trip_failed_then_recover(self) -> None:
-        """TC-IT-WF-002: failed add_transition rolls back; further legitimate add succeeds.
-
-        Pre-state: 3-stage chain s0→s1→s2 (s2 sink). Try to add a duplicate
-        APPROVED edge s0→s1, which trips ``transition_duplicate``. The
-        original Workflow is unchanged. Then add a REJECTED back-edge s1→s0,
-        which is unique on (s1, REJECTED) and preserves s2 as the sole sink.
-        """
-        s0 = make_stage()
-        s1 = make_stage()
-        s2 = make_stage()
-        e0 = make_transition(
-            from_stage_id=s0.id, to_stage_id=s1.id, condition=TransitionCondition.APPROVED
-        )
-        e1 = make_transition(
-            from_stage_id=s1.id, to_stage_id=s2.id, condition=TransitionCondition.APPROVED
-        )
-        wf = make_workflow(stages=[s0, s1, s2], transitions=[e0, e1], entry_stage_id=s0.id)
-
-        bad_dup = make_transition(
-            from_stage_id=s0.id, to_stage_id=s1.id, condition=TransitionCondition.APPROVED
-        )
-        with pytest.raises(WorkflowInvariantViolation):
-            wf.add_transition(bad_dup)
-        assert len(wf.transitions) == 2  # original unchanged
-
-        e_back = make_transition(
-            from_stage_id=s1.id, to_stage_id=s0.id, condition=TransitionCondition.REJECTED
-        )
-        updated = wf.add_transition(e_back)
-        assert len(updated.transitions) == 3
-
-        # remove_stage(s2): cascades the s1→s2 edge; the s1→s0 REJECTED back
-        # edge survives, leaving s1 as the sink (no outgoing from s1 after the
-        # cascade keeps the s1→s0 only). Wait: s1→s0 REJECTED keeps s1 with
-        # outgoing. s0 has outgoing to s1. So no sink. We can't safely
-        # remove s2 here. Instead, demonstrate stage_not_found rollback.
-        with pytest.raises(WorkflowInvariantViolation):
-            updated.remove_stage(uuid4())
-        # After failed remove, original updated stays unchanged.
-        assert len(updated.stages) == 3
-
-    def test_t1_payload_variants_all_rejected(self) -> None:
-        """TC-IT-WF-003: bad payload variants (role / uuid / missing entry / dup stage) reject."""
-        # Variant 1: bad role.
-        v1 = build_v_model_payload()
-        stages_v1 = cast("list[dict[str, object]]", v1["stages"])
-        stages_v1[0]["required_role"] = ["UNKNOWN_ROLE"]
-        with pytest.raises((ValidationError, WorkflowInvariantViolation)):
-            Workflow.from_dict(v1)
-
-        # Variant 2: bad UUID.
-        v2 = build_v_model_payload()
-        v2["id"] = "not-a-uuid"
-        with pytest.raises((ValidationError, WorkflowInvariantViolation)):
-            Workflow.from_dict(v2)
-
-        # Variant 3: missing entry_stage_id.
-        v3 = build_v_model_payload()
-        del v3["entry_stage_id"]
-        with pytest.raises((ValidationError, WorkflowInvariantViolation)):
-            Workflow.from_dict(v3)
-
-        # Variant 4: duplicate stage_id.
-        v4 = build_v_model_payload()
-        stages_v4 = cast("list[dict[str, object]]", v4["stages"])
-        # Set the second stage's id equal to the first stage's id.
-        stages_v4[1]["id"] = stages_v4[0]["id"]
-        with pytest.raises((ValidationError, WorkflowInvariantViolation)):
-            Workflow.from_dict(v4)
